+++ addOns/graphql/src/main/java/org/zaproxy/addon/graphql/ExtensionGraphQl.java
@@ -22,6 +22,7 @@ package org.zaproxy.addon.graphql;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.security.InvalidParameterException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -189,7 +190,7 @@ public class ExtensionGraphQl extends ExtensionAdaptor
         if (extScript != null && extScript.getScript(scriptName) == null) {
             ScriptType variantType =
                     extScript.getScriptType(ExtensionActiveScan.SCRIPT_TYPE_VARIANT);
-            ScriptEngineWrapper engine = extScript.getEngineWrapper("Oracle Nashorn");
+            ScriptEngineWrapper engine = getEngine(extScript, "Oracle Nashorn");
             if (variantType != null && engine != null) {
                 File scriptPath =
                         Paths.get(
@@ -212,6 +213,15 @@ public class ExtensionGraphQl extends ExtensionAdaptor
                 extScript.addScript(script, false);
             }
         }
+    }
+
+    private static ScriptEngineWrapper getEngine(ExtensionScript ext, String engineName) {
+        try {
+            return ext.getEngineWrapper(engineName);
+        } catch (InvalidParameterException e) {
+            LOG.warn("The {} engine was not found, script variant will not be added.", engineName);
+        }
+        return null;
     }
 
     @Override