+++ addOns/soap/src/main/java/org/zaproxy/zap/extension/soap/ExtensionImportWSDL.java
@@ -23,6 +23,7 @@ import java.awt.event.KeyEvent;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.security.InvalidParameterException;
 import javax.swing.JFileChooser;
 import javax.swing.SwingUtilities;
 import javax.swing.filechooser.FileNameExtensionFilter;
@@ -191,7 +192,7 @@ public class ExtensionImportWSDL extends ExtensionAdaptor {
         if (extScript != null && extScript.getScript(scriptName) == null) {
             ScriptType variantType =
                     extScript.getScriptType(ExtensionActiveScan.SCRIPT_TYPE_VARIANT);
-            ScriptEngineWrapper engine = extScript.getEngineWrapper("Oracle Nashorn");
+            ScriptEngineWrapper engine = getEngine(extScript, "Oracle Nashorn");
             if (variantType != null && engine != null) {
                 File scriptPath =
                         Paths.get(
@@ -214,6 +215,18 @@ public class ExtensionImportWSDL extends ExtensionAdaptor {
                 extScript.addScript(script, false);
             }
         }
+    }
+
+    private static ScriptEngineWrapper getEngine(ExtensionScript ext, String engineName) {
+        try {
+            return ext.getEngineWrapper(engineName);
+        } catch (InvalidParameterException e) {
+            LOG.warn(
+                    "The "
+                            + engineName
+                            + " engine was not found, script variant will not be added.");
+        }
+        return null;
     }
 
     @Override